Set better source location for new wanted constraints
It seems TcPluginM's `newWanted` constraint throws away the source
location of the `CtLoc` it is given. Meaning all new [W]anted
contraint get a source-location equal to line 1, column 1.

For all new [W]anted constraints we generate, we set the source
location to be equal to the [W]anted constraint that the solver
is currently trying to solve.